Ignore listing for exceptions considers functions exception will hit
We want the ignore listing logic for pause on caught/uncaught
exception to consider those functions an exception/rejection is
predicted to hit, whether synchronous or asynchronous. Tests
illustrate changed behavior on stopping on exceptions in
ignore listed code, particularly caught exceptions, where we
previously would not stop if the top frame is ignored but will now
stop if any frame it is expected to pass through is not ignored,
and also asynchronous rejections, where we previously only
considered the synchronous stack frames.

Fixed: chromium:1291064
Change-Id: I985808a1a338012c9733d60b5ec7667288948b97