[wasm] Add use counter for module compilation
In addition to the feature-specific use counters, this adds a use
counter that is incremented for every compiled module. This will allow
us to better estimate the percentage of modules that have certain
features enabled.

Since setting a single use counter has some overhead (like creating a
HandleScope) we also change the API to allow to set multiple use
counters at once.

R=mlippautz@chromium.org, jkummerow@chromium.org

Change-Id: I69afa2bbec8fe8bed1ee1037fec39ec5f773a04b