[test] Skip large memory tests in Windows debug mode
`huge-typedarrays` and `wasm/memory64` tests allocate large buffers.
Even though the memory allocation is committed, usually OS does not
actually allocate them until accessed, so it's fine. But the behavior is
different in Windows debug mode. The memory is actually allocated in
Windows debug mode even if they are not accessed.

This makes infra or other tests fail frequently, so skip them in Windows
debug mode.

Bug: v8:4153
Change-Id: I15eaaa8c0484e7d6e4db92083431380f009becc1