[wasm-gc][compiler] load elimination: Fix deadness propagation
This is a follow-up to crrev.com/c/4905772 .
It replaces a node with dead() which is opcode kDead. However
it seems that this can result in ending up with kDeadValue operations
as a result which are not expected by the load elimination.

So if we load from a value that is unreachable, we don't only replace
it with the dead value but also treat the control flow as unreachable
and merge the graph to the end node.

Fixed: chromium:1491415
Change-Id: Ida4018c9c6229acdfe70cffe905890c2358a3032