[heap] Ignore current LAB in SemiSpaceNewSpace::Size
This method is mostly used during GC where we anyways don't have
a LAB in new space. Other use cases are in the allocation slow path
where it shouldn't change much since the LAB is already too small
for the current object.

Drive-by: Also remove the LAB access in
PagedSpaceForNewSpace::AllocatedSinceLastGC(). There is never a LAB
set up when invoking this method.

Bug: chromium:1480975
Change-Id: I5ef0fa782fbd65ab19ee37cd1e74d5b027bc2ff4