Bugfix: LinuxPerfBasicLogger should log JS functions
This patch fixes that typo that was introduced in commit
c51041f45400928cd64fbc8f389c0dd0dd15f82f /
https://chromium-review.googlesource.com/c/v8/v8/+/2336793, which reversed the
behavior of the perf_basic_prof_only_functions flag.

This also refactors the equivalent guard in LinuxPerfJitLogger to use the same
inline CodeKind API for identifying JS Functions. This is unrelated to the bug,
but it seems a fair rider to add on here.

This fixes https://bugs.chromium.org/p/v8/issues/detail?id=14387.