[wasm] Ignore import wrappers if generic wrapper enabled
Just iterating all imports and checking if the generic wrapper can be
used creates overhead. Avoid this if the generic wrapper is generally
enabled. This makes the fast path faster, at the expense of potentially
skipping compilation of import wrappers during compile time - they will
then be compiled later at instantiation time (similarly to when a module
is imported into a new isolate and that isolate does not have the import
wrappers yet).

As the goal is to be able to use import wrappers for the vast majority
of cases, this should be a win overall

R=ahaas@chromium.org

Change-Id: I37cbcbfc673fffc78d8884cb8e2e788ae550f278