[wasm][turboshaft] WasmGCTypeReducer: Resolve aliases for type data
When having chains of aliasing nodes, the "original" operation's type
needs to be updated as well.

Example:
  local.get 0
  ref.cast anyref
  ref.cast structref
  local.get 0
  ref.cast structref

Local[0] will be typed anyref after the first cast. Then the second
cast previously would update the known type to structref for the cast
result of anyref. It did not update the type of local.get 0 to be
structref as well so the last cast could not be removed.

With this change, types are only tracked for non-aliasing nodes meaning
ref.cast and ref.as_non_null do not type their result but only update
the type of their input.
This requires all type lookups to resolve these aliases.

Bug: v8:14108
Change-Id: I9cd62a1ebcc6040ef6d8f29a315f4f0414be6a14