Always use CompressObject when compressing tagged slots
This CL removes the special case for compressing MaybeObject slots,
and now also uses CompressObject for them instead of CompressAny. The
effect of that is that we can also assume that the cage base matches,
and in debug builds we have a DCHECK for that. This is beneficial when
moving objects into trusted space (and out of the main pointer
compression cage) as it immediately catches any remaining tagged
(instead of indirect) pointers to moved-out objects.

Bug: chromium:1473677
Change-Id: Id93c5d389752352ad0347628916c78139def89eb