[heap] Fix broken DCHECK in %SimulateNewspaceFull()
The code path that created a small filler object at the end of a
page, didn't increment allocation_top(). But AddFreshPage() was
DCHECKing that allocation_top() was at the end of the page.

This CL fixes this by introducing a new method
FillCurrentPageForTesting() which puts a filler object at the end of
the page and increments allocation_top() as well. In order for this
to work, we need to free the LAB before this operation. This CL
does this by freeing the LAB after each individual operation for
simplicity. Each new allocation can assume that no LAB exists.

Bug: chromium:1494924
Change-Id: Id5c43432886fef180a22abe647486633b6492229