[debug] Reimplement TemporaryObjectsTracker to fix remaining issues
TemporaryObjectsTracker didn't handle the cases correctly when new
allocation regions overwrite multiple existing regions (e.g. because a
large object is allocated on top of multiple small dead regions).

This CL is a full rewrite of the tracker to not just properly handle
these cases, but also do so in a somewhat clean implementation.

Bug: v8:14420
Change-Id: Ifeda1f0b727a382b339f3b39ef4f62e97f1eada9