[builtins] Unify how Promise.all and .allSettled track resolution
Promise.all had a clever hack to track whether a resolve closure was
already called, but that didn't work when .allSettled needed to track
the state of a promise across both resolve and reject. Tracking was
changed to checking a hole in the values array. This CL now migrates
Promise.all to the same mechanism; avoiding the need to check the old
mechanism.

Change-Id: I501a660659804c27f16f8676f55f94953a6b089b