[wasm] Ignore empty recursion groups in canonicalization
We currently add them to the set of canonical groups, which probably
does not hurt, but creates the strange situation that we have multiple
entries mapping to the same canonical ID.

This CL fixes that by just ignoring empty groups instead of adding them
to the set of canonical groups. Looking up the canonical ID of empty
groups is forbidden anyway.

R=jkummerow@chromium.org

Change-Id: Ia72af5c29b6ea3d2d600602d817d1d765405c0bb