[heap] Fix crash in Scanveger new space allocations
RuntimeCallStats should only be used on the main thread. During
Scavenger cycles, new space labs can be allocated on background threads.
a RCS scope was established in such cases (even though they should not
exist) because the regular new space MainAllocator is used (which is not
marked as a compaction space).

Change-Id: I284ceaf403c8bd97d4b0d3a0008c35cf13c11845