[heap] Avoid usage of CompactionSpaceKind in MainAllocator
Instead of Allocator::GetCompactionSpaceKind() the constructor gets
an additional MainAllocator::Context argument to decide whether this
is an allocation during GC or not. This is similar to the
ConcurrentAllocator usage in a GC and it also will allow us to
get rid of CompactionSpaces in the future.

In addition this CL also allocation_counter_ and
linear_area_original_data_ optional. These fields are not needed
during a GC and don't need to be set up at all.

Bug: chromium:1480975
Change-Id: I06681942c530a88b1be94cbf4327188dc591ad3f