[maglev] Hoist some phi input untagging out of loops
Implement the ability to hoist untagging operations for untagged loop
phis out of the loop.

If we can hoist the untagging we ignore the input edge in
representation inference, since it should always be cheaper to untag
before the loop.

For now the optimization is only enabled in OSR code. The reason is
that in normal code we could have a deopt loop if we can't learn when
the speculative untagging failed.

To be able to insert checked untaggings we create a use a new
checkpointed jump instruction that has a deopt frame attached.

Bug: v8:7700
Change-Id: Ic3f3b0809374ffad412ef9028bb047dd6cd57ad0