Fix store order in SharedFunctionInfo::SetData
Since the SFI's data is accessed from background threads but is now
implemented as two fields (function_data and trusted_function_data), the
order that we store to these fields matters: we need to first store the
new value, then potentially clear the old value (if it was stored in the
other field).
In the future, we'll want to disallow transitions from function_data to
trusted_function_data and vice versa to entirely avoid these kinds of
issues. For that, we'll need to reference Code and InterpreterData also
through the trusted_function_data field.

Bug: v8:14452
Change-Id: Ic42996e4aef2b1ea9cf2729020a43f17734e891f