[heap] Merge StartIncrementalMarkingIfLimitIsReached for all threads
Merge this method method for main and background threads. For this
it now gets the current LocalHeap as an argument.

Note that this changes behavior in ConcurrentAllocator. So far it
always used the background thread version of this method which
never started incremental marking right away. However, with this
change that might happen now.

This means that we now need to start incremental marking before
allocation like we already do on the main thread in order for
black allocation to work.

Bug: chromium:1480975
Change-Id: I02460e07fc159a443e5e2ba00e48e988707cdcbb