Add reason and kind to V8 deoptimization trace event.
There are many reasons for code to be deoptimized in V8. Currently there
is a trace event that just notifies about the deoptimization but does
not provide any hints on why the code has been deoptimized. This CL
proposes adding some info to trace event to make it a little bit more
informative.

Change-Id: I170daff9822ce19c0e537ecc55fdf9180135a475