Verify that thought-to-be dead code is indeed dead
Now that we convert the BytecodeWrapper instead of the BytecodeArray
into UncompiledData during bytecode flushing (see crrev.com/c/4993368),
we can no longer run into the "bytecode already decompiled" case, so it
should be dead code. This CL first inserts a CHECK to verify that it is
indeed dead, and a subsequent CL will then remove the logic.

Bug: chromium:1472252
Change-Id: If83606d041fb8b92ef0f6c5136a5b04ae0a59b32