[sandbox] Introduce TrustedFixedArray
This is similar to a FixedArray, except that it lives in trusted space.
This will for example be used for DeoptimizationData. For now, any
compressed pointer stored in a TrustedFixedArray uses the main pointer
compression cage base, but this will also likely change in the future.
It may also be the case that we need two variants of these arrays: one
that uses the main cage base (e.g. for constant pools) and one that uses
the trusted cage base (e.g. for deoptimization data).

Bug: chromium:1473677
Change-Id: Iebf018f041c83d900b53eec572dbcc2fe286c6be