[heap] Use HeapAllocator in LocalHeaps for managing LABs
HeapAllocator is the class that manages all LABs for the main thread.
However, background threads so far had all their LABs as fields on
LocalHeap. With this CL LocalHeap now also manages its LABs with
the HeapAllocator class.

The main thread had potentially two separate LABs in the old space.
One in Heap::heap_allocator_ and one through its LocalHeap. This CL
also removes this by making Heap::heap_allocator_ a pointer to the
HeapAllocator in LocalHeap.

Note that LocalHeap still uses a separate implementation for
LocalHeap::AllocateRawWith. We can also merge this method with
Heap::AllocateRawWith as a follow-up.

Bug: chromium:1480975
Change-Id: I3d4c75a6802151b52b41bd5cbf5dbcc954d4c28c