[runtime] optimize dictionary mode key enumeration
When enumerating large dictionary mode objects, we need to sort their
properties by their dictionary index, which is slow. The sorting
currently uses the comparison-based std::sort. The dictionary indexes
are distinct, mostly contiguous, and small integers. This makes them
very amenable to non-comparison sorts.

This CL implements pigeonhole sorting. First a big array is allocated
with `max_dictionary_index - min_dictionary_index + 1` elements. Then
each enumerable string property is placed at `dictionary_index -
min_dictionary_index`. Since dictionary indexes are non-contiguous, this
process may leave holes in the array, so a second pass may be needed to
compact the array and right-trim the suffix.

The problem with pigeonhole sorting is that it's slow in the (unlikely?)
case where the user inserts a bunch of keys into an object, and then
deletes almost all of the keys in the middle. In that case the range of
valid dictionary indexes can be arbitrarily larger than the number of
keys in the object. I chose to handle that by leaving the existing
comparison-sort code in, and adding a heuristic to select between the
two methods.

I used the benchmark in the attached bug to compare performance:

N = 1:    identical.
N = 5:    1.1x speedup.
N = 10:   1.24x speedup.
N = 100:  2x speedup.
N = 1000: 3.5x speedup.

Similar improvements were observed for for-in performance. This
optimization could also be applied for Object.values/entries, but those
are not done in this CL.

I ran tests locally. I also ensured that neither code branch was dead by
commenting each one out and checking that tests failed.

Bug: v8:12457
Change-Id: I21e04f19a182ce74cf32cfac0b021355a4ead306