[turboshaft] Normalize loads whose base is addition and index is invalid
This make the instruction selector generate better code (because the
addition is computed with a complex memory operand rather than before
the load).

Additionally, this makes the TS graphs more normalized, which could be
good in particular for LoadElimination and StoreStoreElimination, and
any alias analysis we may want to do.

Bug: v8:12783
Change-Id: I08f75d83c972f2fd60c064fce2ab944f0541f7ec