[sandbox] Use zero-extension when bounds checking in AccessorAssembler
In EmitFastElementsBoundsCheck, we were previously using SmiUntag when
loading the length of the JSArray, which will potentially perform a sign
extension of the length. Normally, this doesn't matter as the length of
an array will always be a positive Smi value. However, given the sandbox
attacker model, this no longer holds: an attacker could change the
length to be a negative value, in which case the following bounds check
would be wrong. This could allow an attacker to reach out of the sandbox
with a very large negative index. This CL now fixes this by using
PositiveSmiUntag instead, which will perform zero extension.

Bug: chromium:1507691
Change-Id: I85a296d34686bb08fa97c1d71af6ddea543a84b5