[sandbox] Introduce SBXCHECK
When the sandbox is enabled, a SBXCHECK behaves exactly like a CHECK,
i.e. will terminate the process in both debug and release configurations
if the condition is not true. However, semantically it indicates that
the check is required for the sandbox, i.e. prevents a (potential)
sandbox bypass. When the sandbox is off, a SBXCHECK becomes a DCHECK.

Change-Id: I9e46dd2efcb733f596137c4107ef8fe3f02ff73c