[turboshaft][wasm] Resolve Phi of RttCanons
Under specific circumstances involving a combination of the
ValueNumberingReducer and loop unrolling, we ended up with a phi as
an rtt input for a wasm-gc allocation. This was not handled correctly
by the `WasmGCTypeReducer`, which expected an `RttCanonOp` input.
We are fixing this by resolving phis to identical rtts in the
RequiredOptimizationReducer.

Bug: chromium:1507779
Change-Id: Iacafc6982ec2a22655203a47cbb8d3f0a6785058