Update patch set 4
Patch Set 4:

This CL has failed the run. Reason:

Failed Tryjobs:
* [v8/try/v8_linux64_header_includes_dbg](https://cr-buildbucket.appspot.com/build/8762508609419178177). Summary [(view all results)](https://chromium-review.googlesource.com/c/5077500?checksPatchset=4&tab=checks):

---
#### Step _generate_build_files_ failed. Error logs are shown below:
```
ERROR at //src/compiler/code-assembler.cc:10:11: Include not allowed.
#include "src/codegen/code-stub-assembler-inl.h"
          ^------------------------------------
It is not in any dependency of
  //:v8_compiler_for_mksnapshot_source_set
The include file is in the target(s):
  //:v8_initializers
which should somehow be reachable.
___________________
ERROR at //src/compiler/code-assembler.cc:10:11: Include not allowed.
#include "src/codegen/code-stub-assembler-inl.h"
          ^------------------------------------
It is not in any dependency of
  //:v8_compiler
The include file is in the target(s):
  //:v8_initializers
which should somehow be reachable.
```
#### More information can be found in the stdout.

---
* [v8/try/v8_linux_arm_lite_compile_dbg](https://cr-buildbucket.appspot.com/build/8762508609419177985). Summary [(view all results)](https://chromium-review.googlesource.com/c/5077500?checksPatchset=4&tab=checks):

---
#### Step _generate_build_files_ failed. Error logs are shown below:
```
ERROR at //src/compiler/code-assembler.cc:10:11: Include not allowed.
#include "src/codegen/code-stub-assembler-inl.h"
          ^------------------------------------
It is not in any dependency of
  //:v8_compiler_for_mksnapshot_source_set
The include file is in the target(s):
  //:v8_initializers
which should somehow be reachable.
```
#### More information can be found in the stdout.

---

Patch-set: 4
Tag: autogenerated:cq:dry-run:1701817609
Attention: {"person_ident":"Gerrit User 1183889 \u003c1183889@3ce6091f-6c88-37e8-8c75-72f92ae8dfba\u003e","operation":"ADD","reason":"ps#4: Tryjobs failed"}