[turboshaft][cleanup] Use more ZoneVectors
... in `UseMap` and `ComputeSpecialRPO`.

Since the original `std::vectors` were allocated just once per function,
this probably has no effect on compile time, but it cleans up the
heaptrack profile a bit, which is helpful for future investigations.

Concretely, this change reduces the number of malloc calls for `UseMap`
from ~102k to 25k (-75%) and for `ComputeSpecialRPO` from 624k to 206k
(-66%), measured locally on an x64 workstation with heaptrack, while
eagerly compiling a large Wasm module (PSWeb, ~50 MB of code) with
Turboshaft.

Change-Id: I397962c3ce526cd48e07fc51a7b1ebe3f386bb55