[heap] Simplify PagedSpaceAllocatorPolicy::ContributeToSweepingMain
This method was using a lot of arguments that were actually not
really required:

1. The arguments `origin` and `size_in_bytes` were only needed because
this method was also allocating. The allocation is now tried again
after ContributeToSweeping() in RefillLab().
2. `sweeping_scope_id` and `sweeping_scope_kind` were computed in
PagedSpaceAllocatorPolicy::RefillLab and then passed to
ContributeToSweeping(). It was only needed in RefillLab() for
RefillFreeList(). The scope for RefillFreeList() alone doesn't seem
warranted and we can move that code into ContributeToSweeping().
3. `required_freed_bytes` was passed to ContributeToSweeping() but
wasn't really used. Sweeping anyways stopped after a single page and
was unconstrained in the second invocation. However, we anyways
did not check the result of ParallelSweepSpace() before retrying the
allocation. It is simpler to simply retry the allocation. Concurrent
allocation anyways means that the memory could be "stolen" by some
other thread between sweeping and retrying the allocation.

Bug: chromium:1480975
Change-Id: I23ef37724e6ba364870e0ca8c64d81c086d737a4