[heap] Simplify PagedSpaceAllocatorPolicy::ContributeToSweepingMain
This method was using a lot of arguments that were actually not
really required:

1. The arguments `origin` and `size_in_bytes` were only needed because
this method was also allocating. We now just invoke
TryAllocationFromFreeList() again after invoking ContributeToSweeping().
2. `sweeping_scope_id` and `sweeping_scope_kind` were computed in
PagedSpaceAllocatorPolicy::RefillLab and then passed to
ContributeToSweeping(). It was only needed in RefillLab() for
RefillFreeList(). The scope for RefillFreeList() alone doesn't seem
warranted and we can move that code into ContributeToSweeping().
3. `required_freed_bytes` was passed to ContributeToSweeping() but
wasn't really used. Sweeping anyways stopped after a single page and
was unconstrained in the second invocation. However, we anyways
did not check the result of ParallelSweepSpace() before retrying the
allocation. We simply unconditionally retry the allocation. Concurrent allocation anyways means that the memory could be "stolen"
by some other thread between sweeping and retrying the allocation.

MainAllocator was the only user of the `required_free_bytes` argument
for ParallelSweepSpace. All other uses passed 0 to this argument
and left this counter unconstrained. Removing this argument also lets
us remove the corresponding counter in RawSweep(). It was also the
only user of FreeList::GuaranteedAllocatable() which could now be
removed.

Drive-by: Remove the `Main` suffix for methods in
PagedSpaceAllocatorPolicy. This suffix used to signal that this was
used from the main thread but since this code now runs on
background threads as well it is outdated.

Bug: chromium:1480975
Change-Id: I23ef37724e6ba364870e0ca8c64d81c086d737a4