[turboshaft] Make Assembler more generic
This CL is part of an effort to make the reducer framework more
generic so that it can be used with Maglev as well.

The only reducer of the stack that is required but not generic is now
EmitProjectionReducer; a subsequent CL will address this (but this is
a bit trickier, since it needs both technical insights into the IR,
but also needs to work for both Turboshaft and Maglev IR).

Bug: v8:12783
Change-Id: I9c3c25e3e4cce00228d87025b473d59e488d1890