Restrict bit_cast
Using bit_cast on pointers does not accomplish anything useful, and
obscures what might be incorrect behavior. Do not allow base::bit_cast
to be misused in that way.

Port the checks present in https://crrev.com/c/5094596 to this repo's
version of bit_cast. Leave the implementation that uses memcpy in
place, as V8 (unlike Chromium) cannot necessarily rely on compiler
intrinsics.

Bug: chromium:1506769
Change-Id: Ie05d68c3e909038de9406a05a0b9109f3509e08b