[multimodal] Multimodal tiering decisions
This is a first step towards multimodal tiering decisions.

We introduce two tunables. One is efficiency, where we compromise
throughput for latency. The other is battery saving, where we optimize
for total cpu cycles spent (or in theory energy used).

These two settings can be set by the embedder. Currently chrome already
marks background isolates, therefore we do not yet couple the background
notification with the efficiency mode, but rather put that behind the
`--efficiency_mode_for_background_isolate` flag, which allows us to do a
controlled experiment.

The main direction logic of this CL is:

* To be efficient: use synchronous compilation
* To save battery: deferr tier-up to TF

Bug: v8:14515
Change-Id: If9b4d2f0f14586eaa4174abc485db21673625f8d