[wasm][fuzzer] Disable lazy compilation
The wasm fuzzer takes the input bytes and just feeds them to Wasm module
compilation. If
1) that succeeds,
2) there is no start function, and
3) there is an export named "main"
it will run that export.

As it's unlikely to randomly generate a module which satisfies all these
conditions, the fuzzer mostly tests compilation.

Hence disable lazy compilation, and choose one of the two compilers
(Liftoff or TurboFan) for the whole module. This increases compiler
coverage of the fuzzer.

R=ahaas@chromium.org, mliedtke@chromium.org

Change-Id: Ic7ce1fd7a24bf0eb7047fda3351391f832e11d36