[wasm][compiler] WasmLoadElimination: Skip struct.set with None() type
This is a follow-up to commit 949bbdf4c2583faf581d6e4b67854c0e0200dd5b.
It repeats the same fix for struct.set that was done for struct.get.
I also added a comment that explains how we can potentially end up in
that situation.

We also have already precedent for checking for kDeadValue in
commit 9ece4a12102e77976aeb8f5b1879e93b700acaac and
commit c0cd50fc1d562ef7a2eced5a63504d3b776c82de among others.
As a more long-term solution we should probably adapt all checks
for kDead in these reducers (WasmLoadElimination and
WasmGCOperatorReducer to always also check for kDeadValue.

Bug: chromium:1507106
Change-Id: Ieed41c927dbcf199df2c92d4508c7fe67a23f05c