[multimodal] Multimodal tiering decisions
This is a first step towards multimodal tiering decisions.

We introduce two tunables. One is efficiency, where we compromise
throughput for latency. The other is battery saving, where we optimize
for total cpu cycles spent (or in theory energy used).

These two settings can be influenced by the embedder. Battery mode is
decided solely by the embedder. For efficiency mode we currently plan
to consider the background setting of an isolate, but might add more
signals in the future.

Since chrome already marks background isolates the coupling is opt-in
for now, controlled by `--efficiency-mode-for-background-isolate`.

The gist of the current adjustments are:

* To be efficient: use synchronous compilation, optionally delay or
  defer tier-up
* To save battery: defer tier-up to TF

In either battery or efficiency mode the only OSR enabled is from
unoptimized code to maglev.

Bug: v8:14515
Change-Id: If9b4d2f0f14586eaa4174abc485db21673625f8d