[liftoff] Enable root-relative access
This sets the enable_root_relative_access assembler option. For Liftoff
code it is totally fine to emit root-relative accesses, and we do so
explicitly in some places.

The consequences of setting the option are a bit intransparent though,
so land it in isolation so we can see any regressions or crashes
introduced by this.

R=ahaas@chromium.org

Bug: v8:14499
Change-Id: Ifcfaa4941d07ed1bdad60d9a3df80ec22f8f5b8d