[builtins] Track topmost script-having execution context
The topmost script-having execution context is defined by the Web IDL
spec here:
https://html.spec.whatwg.org/multipage/webappapis.html#topmost-script-having-execution-context)
In short it's the context of the topmost user JavaScript code if
available and it's used in the definition of the incumbent context
(see https://html.spec.whatwg.org/multipage/webappapis.html#incumbent-settings-object).

This CL adds machinery for tracking topmost script-having execution
context when it's known "for free" which allows to speed up
v8::Isolate::GetIncumbentContext() for those cases.

Bug: v8:13825
Change-Id: I52d137eabb91962bdde389018bcfa9aada3fde4a