[maglev] Cache TypedArray length loads
Non-resizable TypedArrays can't change length, so we can cache loads of
their length (for bounds checks) as a constant property.

Since properties are cached by name, and users can overwrite "length"
on a TypedArray, this additionally introduces a mechanism for using
a non-Name value as the key into the properties cache.

Bug: v8:7700
Change-Id: I59e74ed7fdc5929d2de7146ee4d36b14ef35808e