Handle OOM in VirtualAddressSubspace::FreePages
It can happen that the decommitting of the underlying pages fails, for
example when running out of VMAs. In that case, instead of a CHECK
failure we should fail with an out-of-memory error. While this would
normally happen via FatalProcessOutOfMemory, this function is not
available in src/base. To deal with this, this CL introduces a simpler
version: base::FatalOOM, which fails with a similar error message so
that fuzzers and other tooling can detect these crashes as OOMs.

Bug: chromium:1515110
Change-Id: I10f2c6b63a35e7d54c9cd4e548c0e0eb1fd2c61f