Introduce base::FatalOOM
This function is a simpler version of FatalProcessOutOfMemory that
simply terminates the process with a recognizable message. It is now
used as fallback for the higher level OOM routines, but also directly
from src/base and a few other places where those are not available.

In addition, the new FatalOOM function is now used in
VirtualAddressSpace::FreePages as it can happen that the decommitting of
the underlying pages fails, for example when running out of VMAs. In
that case, instead of a CHECK failure we should fail with an
out-of-memory error.

Bug: chromium:1515110
Change-Id: I10f2c6b63a35e7d54c9cd4e548c0e0eb1fd2c61f