[wasm][ia32] Fix atomic binops with constants
https://crrev.com/c/5180736 added support for atomic operations at
constant memory locations. On ia32 we had an optimization which swapped
the memory start register and the offset register under certain
conditions to save register moves. Later code then only handled the case
that the offset register is not set, and not that the destination
register is not set.

We hence move the check whether to swap the registers to the branch that
handles the case that both registers are actually there.

R=ahaas@chromium.org

Fixed: chromium:1519091
Change-Id: Ic68a5562a0894e021ca26c7d0fd87c6b609a2a56