Make CrossScriptInliningCallerLineNumbers2 test less flaky
Edge is observing this test flaky. As the inner loop became faster,
the cpu profiler samples execution of `action` with a less chance, which
resulted in missing `action` child under `level3_node`, therefore the
test crashes.

Make the loop a bit slower so that it gives more chance for the cpu
profiler to collect more samples of the `action` function.

Change-Id: I8c1763bb55f28668e92479119dcb27fe49d97c8a