[maglev] Fix concurrent dispatcher flushing again
Destructing MaglevJob's on a background thread needs an unparked local
isolate and therefore cannot be done during flushing.

Instead we can perform the destruction on the main thread. New
destruction jobs are only produced on the main thread, thus both input
and destruct queues are guaranteed to be empty when we `Join`.

Fixed: v8:14588
Change-Id: Id534266e7f6254f773a0a6246e331ac441d1323a